영속성 전이 스터디 및 리팩토링
영속성 전이, 고아객체 스터디
테스트 구조 리팩토링
