+++ ex1-hello-jpa/src/main/java/hellojpa/Team.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Getter
@@ -20,6 +21,11 @@ public class Team extends BaseEntity {
     private Long id;
     private String name;
 
-    @OneToMany(mappedBy = "team")
-    private List<Member> members;
+    @OneToMany(mappedBy = "team", cascade = CascadeType.REMOVE, orphanRemoval = true)
+    private List<Member> members = new ArrayList<Member>();
+
+    public void addMember(Member member) {
+        members.add(member);
+        member.setTeam(this);
+    }
 }
