+++ ex1-hello-jpa/src/main/java/jpa/JpaFunction.java
@@ -1,0 +1,7 @@
+package jpa;
+
+import jakarta.persistence.EntityManager;
+
+public interface JpaFunction {
+    void execute(EntityManager em);
+}
