+++ ex1-hello-jpa/src/main/java/jpa/JpaTemplate.java
@@ -1,0 +1,27 @@
+package jpa;
+
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityManagerFactory;
+import jakarta.persistence.EntityTransaction;
+import jakarta.persistence.Persistence;
+
+public class JpaTemplate {
+    public void execute(JpaFunction function) {
+        EntityManagerFactory emf = Persistence.createEntityManagerFactory("hello");
+        EntityManager em = emf.createEntityManager();
+
+        EntityTransaction tx = em.getTransaction(); //JPA의 엔티티 EntityTransaction 얻음
+        tx.begin(); //트랜잭션 시작
+
+        try {
+            function.execute(em);
+            tx.commit();
+        } catch (Exception e) {
+            tx.rollback();
+            e.printStackTrace();
+        } finally {
+            em.close();
+            emf.close();
+        }
+    }
+}
