+++ jpashop/src/main/java/jpabook/domain/Album.java
@@ -1,0 +1,17 @@
+package jpabook.domain;
+
+import jakarta.persistence.Entity;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+public class Album extends Item {
+    private String artist;
+    private String etc;
+}
