+++ jpashop/src/main/java/jpabook/domain/Category.java
@@ -21,7 +21,7 @@ public class Category {
     private Long id;
     private String name;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "parent_id")
     private Category parent;
