+++ jpashop/src/main/java/jpabook/domain/Movie.java
@@ -1,0 +1,19 @@
+package jpabook.domain;
+
+import jakarta.persistence.DiscriminatorValue;
+import jakarta.persistence.Entity;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@DiscriminatorValue("M")
+public class Movie extends Item{
+    private String director;
+    private String actor;
+}
