+++ jpashop/src/main/java/jpabook/domain/OrderItem.java
@@ -17,11 +17,11 @@ public class OrderItem {
     @Column(name ="order_item_id")
     private Long id;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "order_id")
     private Order order;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "item_id")
     private Item item;
