+++ jpashop/src/main/java/jpabook/jpa/JpaFunction.java
@@ -1,0 +1,7 @@
+package jpabook.jpa;
+
+import jakarta.persistence.EntityManager;
+
+public interface JpaFunction {
+    void execute(EntityManager em);
+}
