+++ jpashop/src/main/java/jpabook/jpa/JpaTemplate.java
@@ -1,27 +1,24 @@
-package jpabook;
+package jpabook.jpa;
 
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.EntityManagerFactory;
 import jakarta.persistence.EntityTransaction;
 import jakarta.persistence.Persistence;
-import jpabook.domain.Member;
-import jpabook.domain.Order;
-import jpabook.domain.OrderStatus;
 
-import java.util.List;
-
-public class MainTotalTest {
-    public static void main(String[] args) {
-        EntityManagerFactory emf = Persistence.createEntityManagerFactory("jpashop");
+public class JpaTemplate {
+    public void execute(JpaFunction function) {
+        EntityManagerFactory emf = Persistence.createEntityManagerFactory("hello");
         EntityManager em = emf.createEntityManager();
 
         EntityTransaction tx = em.getTransaction(); //JPA의 엔티티 EntityTransaction 얻음
         tx.begin(); //트랜잭션 시작
 
         try {
+            function.execute(em);
             tx.commit();
         } catch (Exception e) {
             tx.rollback();
+            e.printStackTrace();
         } finally {
             em.close();
             emf.close();
